hw&scripts/api_yatube/requirements.txt
pytest-django
requests
pillow
sorl-thumbnail
django
djangorestframework
django-cors-headers
